requirements.txt
Django==1.11.7
django-flat-responsive==2.0
django-recaptcha==1.3.1
pytz==2017.3
